Do not always force async initialization.
Adds a new `async` prop to force async initialization, as proposed in #95.

Async initialization will make your page "jump" when the component appears, as it will first be rendered with the normal size, then resized to content. But async init is only needed when using StyledComponents (see #71) or when `maxRows` is set, so let the user decide to enable it.

Fixes #95